Changed: #825 Remove all warnings when compiling Ryzom

+++ code/nel/src/misc/sheet_id.cpp
@@ -157,7 +157,7 @@ bool CSheetId::buildSheetId(const std::string& sheetName)
 			{
 				_FileExtensions.push_back(sheetType);
 				_DevSheetIdToName.push_back(std::vector<std::string>());
-				typeId = _FileExtensions.size() - 1;
+				typeId = (uint32)_FileExtensions.size() - 1;
 				_DevTypeNameToId[sheetType] = typeId;
 				std::string unknownNewType = std::string("unknown." + sheetType);
 				_DevSheetIdToName[typeId].push_back(unknownNewType);
